add :Z mount flag for temporal container
When running under `podman-compose` and rootless podman on a Linux system with selinux enabled, the `temporal` container does not start because the dynamicconfig mount fails.  This leads to a "500" error page. 

When selinux is enabled, the `Z` flag modifies the selinux label so the podman process can access the bind mounted directory. The option has no effect when running without selinux enabled.

Change was tested with docker 24.0.5 on Unbuntu 23.04 with selinux disabled, and Fedora 38 with podman 4.6 and podman-compose 1.0.6